Reorganize code directory structure: Status, Omni => Components
