+++ src/libraries/JANA/CMakeLists.txt
@@ -62,8 +62,8 @@ set(JANA2_SOURCES
     Services/JServiceLocator.h
     Services/JEventGroupTracker.h
 
-    Status/JComponentSummary.h
-    Status/JComponentSummary.cc
+    Components/JComponentSummary.h
+    Components/JComponentSummary.cc
 
     Streaming/JDiscreteJoin.h
     Streaming/JEventBuilder.h
@@ -197,26 +197,24 @@ endif()
 file(GLOB jana_headers "*.h*")
 file(GLOB jana_engine_headers "Engine/*.h*")
 file(GLOB jana_services_headers "Services/*.h*")
-file(GLOB jana_status_headers "Status/*.h*")
 file(GLOB jana_streaming_headers "Streaming/*.h*")
 file(GLOB jana_utils_headers "Utils/*.h*")
 file(GLOB jana_calibs_headers "Calibrations/*.h*")
 file(GLOB jana_cli_headers "CLI/*.h*")
 file(GLOB jana_compat_headers "Compatibility/*.h*")
 file(GLOB jana_podio_headers "Podio/*.h*")
-file(GLOB jana_omni_headers "Omni/*.h*")
+file(GLOB jana_component_headers "Component/*.h*")
 file(GLOB jana_topology_headers "Topology/*.h*")
 
 install(FILES ${jana_headers} DESTINATION include/JANA)
 install(FILES ${jana_engine_headers} DESTINATION include/JANA/Engine)
 install(FILES ${jana_services_headers} DESTINATION include/JANA/Services)
-install(FILES ${jana_status_headers} DESTINATION include/JANA/Status)
 install(FILES ${jana_streaming_headers} DESTINATION include/JANA/Streaming)
 install(FILES ${jana_utils_headers} DESTINATION include/JANA/Utils)
 install(FILES ${jana_calibs_headers} DESTINATION include/JANA/Calibrations)
 install(FILES ${jana_cli_headers} DESTINATION include/JANA/CLI)
 install(FILES ${jana_compat_headers} DESTINATION include/JANA/Compatibility)
-install(FILES ${jana_omni_headers} DESTINATION include/JANA/Omni)
+install(FILES ${jana_component_headers} DESTINATION include/JANA/Component)
 install(FILES ${jana_topology_headers} DESTINATION include/JANA/Topology)
 
 if (${USE_PODIO})
