+++ src/libraries/JANA/Components/JComponentSummary.cc
