+++ src/libraries/JANA/Components/JOmniFactory.h
@@ -13,7 +13,7 @@
 #include <JANA/JEvent.h>
 #include <JANA/JMultifactory.h>
 #include <JANA/JVersion.h>
-#include <JANA/Omni/JHasInputs.h>
+#include <JANA/Components/JHasInputs.h>
 
 #include <JANA/JLogger.h>
 #include <JANA/Services/JLoggingService.h>
