+++ src/libraries/JANA/JApplicationFwd.h
@@ -25,7 +25,7 @@ class JParameterManager;
 class JApplication;
 extern JApplication* japp;
 
-#include <JANA/Status/JComponentSummary.h>
+#include <JANA/Components/JComponentSummary.h>
 #include <JANA/Engine/JPerfSummary.h>
 #include <JANA/JLogger.h>
